prepare for 3d code
