+++ src/shadowfax/voronoi.h
@@ -1,36 +1,9 @@
-/*******************************************************************************
- * This file is part of cVoronoi.
- * Copyright (c) 2020, 2021 Bert Vandenbroucke (bert.vandenbroucke@gmail.com)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- ******************************************************************************/
+//
+// Created by yuyttenh on 02/07/2021.
+//
 
-/**
- * @file voronoi.h
- *
- * @brief 2D Voronoi grid.
- *
- * @author Bert Vandenbroucke (bert.vandenbroucke@ugent.be)
- */
-
-#ifndef SWIFT_VORONOI_STRUCT_H
-#define SWIFT_VORONOI_STRUCT_H
-
-#include "part.h"
-
-#include <string.h>
+#ifndef SWIFTSIM_VORONOI_H
+#define SWIFTSIM_VORONOI_H
 
 /*! @brief Store the edges of faces (so that the actual Voronoi grid can be
  *  reconstructed). */
@@ -42,92 +15,12 @@
 /*! @brief Store cell generators. */
 #define VORONOI_STORE_GENERATORS
 
-/**
- * @brief Voronoi interface.
- *
- * An interface is a connection between two neighbouring Voronoi cells. It is
- * completely defined by the indices of the generators that generate the two
- * neighbouring cells, a surface area and a midpoint position.
- */
-struct voronoi_pair {
-  /*! Pointer to particle corresponding to the generator on the left of the
-   * interface (always a particle within the local cell). */
-  struct part *left;
-
-  /*! Pointer to particle corresponding to the generator on the right of the
-   * interface (can be a local particle, but also a particle in a
-   * neighbouring cell). */
-  struct part *right;
-
-  struct cell *right_cell;
-
-  /*! Surface area of the interface. */
-  double surface_area;
-
-  /*! Midpoint of the interface. */
-  double midpoint[3];
-
-#ifdef VORONOI_STORE_CONNECTIONS
-  /*! First vertex of the interface. */
-  double a[2];
-
-  /*! Second vertex of the interface. */
-  double b[2];
-#endif
-};
-
-/**
- * @brief Voronoi cell.
- *
- * A cell stores geometrical information about a Voronoi cell: its volume and
- * the location of its centroid.
- */
-struct voronoi_cell_new {
-  /*! Cell volume. */
-  double volume;
-
-  /*! Cell centroid. */
-  double centroid[2];
-
-#ifdef VORONOI_STORE_GENERATORS
-  /*! Position of the cell generator. */
-  double generator[2];
+#if defined(HYDRO_DIMENSION_3D)
+#include "voronoi_3d/voronoi_3d.h"
+#elif defined(HYDRO_DIMENSION_2D)
+#include "voronoi_2d/voronoi_2d.h"
+#else
+#error "Only 2D or 3D moving mesh are supported!"
 #endif
 
-#ifdef VORONOI_STORE_CELL_STATS
-  /*! Number of faces of this cell. */
-  int nface;
-#endif
-};
-
-/**
- * @brief Voronoi grid.
- *
- * The grid stores a copy of the coordinates of the grid generators, the
- * coordinates of the grid vertices and the edge connections that make up the
- * grid. For every generator, it stores the number of vertices for the cell
- * generated by it, and the offset of the cell edges in the edge array.
- */
-struct voronoi {
-
-  /*! @brief Voronoi cells. */
-  struct voronoi_cell_new *cells;
-
-  /*! @brief Number of cells. */
-  int number_of_cells;
-
-  /*! @brief Voronoi cell pairs. We store these per (SWIFT) cell, i.e. pairs[0]
-   *  contains all pairs that are completely contained within this cell, while
-   *  pairs[1] corresponds to pairs crossing the boundary between this cell and
-   *  the cell with coordinates that are lower in all coordinate directions (the
-   *  cell to the left, front, bottom, sid=0), and so on. */
-  struct voronoi_pair *pairs[27];
-
-  /*! @brief Current number of pairs per cell index. */
-  int pair_index[27];
-
-  /*! @brief Allocated number of pairs per cell index. */
-  int pair_size[27];
-};
-
-#endif /* SWIFT_VORONOI_STRUCT_H */
+#endif  // SWIFTSIM_VORONOI_H
