+++ src/shadowfax/voronoi_2d/geometry_2d.h
