+++ src/shadowfax/voronoi_3d/delaunay_functions_3d.h
@@ -1,0 +1,10 @@
+//
+// Created by yuyttenh on 02/07/2021.
+//
+
+#ifndef SWIFTSIM_DELAUNAY_FUNCTIONS_3D_H
+#define SWIFTSIM_DELAUNAY_FUNCTIONS_3D_H
+
+#include "delaunay_3d.h"
+
+#endif  // SWIFTSIM_DELAUNAY_FUNCTIONS_3D_H
