+++ src/shadowfax/voronoi_3d/queues.h
@@ -1,0 +1,8 @@
+//
+// Created by yuyttenh on 02/07/2021.
+//
+
+#ifndef SWIFTSIM_QUEUES_H
+#define SWIFTSIM_QUEUES_H
+
+#endif  // SWIFTSIM_QUEUES_H
