+++ src/shadowfax/voronoi_3d/voronoi_3d.h
@@ -1,0 +1,8 @@
+//
+// Created by yuyttenh on 02/07/2021.
+//
+
+#ifndef SWIFTSIM_VORONOI_3D_H
+#define SWIFTSIM_VORONOI_3D_H
+
+#endif  // SWIFTSIM_VORONOI_3D_H
